+++ examples/showcase/calculate_sum.rs
@@ -3092,7 +3092,7 @@ pub fn calculate_sum(numbers: &Vec<i32>) -> i32 {
     let mut total: i32 = Default::default();
     total = 0;
     for n in numbers.iter().cloned() {
-        total = (total).py_add(n);
+        total = ((total).py_add(n)) as i32;
     }
     total
 }
